third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-tables/width-distribution/computing-column-measure-1-expected.txt

This is a testharness.js-based test.
PASS Checking intermediate min-content height for span 2 (1)
FAIL Checking intermediate min-content height for span 2 (2) assert_equals: expected 10 but got 51
PASS Checking intermediate min-content height for span 2 (3)
FAIL Checking intermediate min-content height for span 2 (4) assert_equals: expected 17 but got 10
Harness: the test ran to completion.
